Have the controller recognize only one command at a time.
It will remember every button-press and store them in an array, but it
will only act on one at a time. Therefore, if multiple buttons are
press, the most recent one is used for the movement command. When that
is released, the second-most recent one is used. And so on.